visualiser is ready

+++ src/main/Grass.java
@@ -2,9 +2,9 @@ package main;
 
 public class Grass implements IMapElement {
     Vector2D position;
-    Integer plantEnergy;
+    Double plantEnergy;
 
-    public Grass(Vector2D position, Integer plantEnergy){
+    public Grass(Vector2D position, Double plantEnergy){
         this.position=position;
         this.plantEnergy=plantEnergy;
     }
